fix: [Refactoring] clean up the main App file
