+++ src/domain/home/text_panel/services_details/MoreDetailsModal.tsx
@@ -1,4 +1,4 @@
-import { GlobalContext } from '@components/context/GlobalContext';
+import { GlobalContext } from '@context/GlobalContext';
 import { Dialog, Transition } from '@headlessui/react';
 import { EnvelopeIcon, XMarkIcon } from '@heroicons/react/24/outline';
 import React, { Fragment, useContext } from 'react';
